requirements.txt
arabic-reshaper==3.0.0
argcomplete==3.1.2
asgiref==3.7.2
asn1crypto==1.5.1
aws-psycopg2==1.3.8
awscli==1.29.66
bootstrap4==0.1.0
boto3==1.28.66
botocore==1.31.66
certifi==2023.7.22
cffi==1.16.0
cfn-flip==1.3.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.4
crispy-bootstrap4==2023.1
cryptography==41.0.4
cssselect2==0.7.0
dj-database-url==2.1.0
Django==4.2.6
django-ajax-selects==2.2.0
django-crispy-forms==2.1
docutils==0.16
durationpy==0.5
gunicorn==21.2.0
hjson==3.1.0
html5lib==1.1
idna==3.4
jmespath==1.0.1
kappa==0.6.0
lxml==4.9.3
MarkupSafe==2.1.3
oscrypto==1.3.0
packaging==23.2
Pillow==10.1.0
placebo==0.9.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pypdf==3.16.4
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
rsa==4.7.2
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
style==1.1.0
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
tqdm==4.66.1
troposphere==4.4.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
update==0.0.1
uritools==4.0.2
urllib3==2.0.6
webencodings==0.5.1
Werkzeug==3.0.0
whitenoise==6.6.0
xhtml2pdf==0.2.11
